pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "headless-chat"
version = "0.1.0"
description = "Headless Chat API with FastAPI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "sqlalchemy",
    "asyncpg",
    "redis",
    "pydantic-settings",
    "alembic",
    "python-multipart",
    "python-jose[cryptography]",
    "passlib[argon2]",
    "aiosqlite",
    "aiofiles",
    "pytest",
    "pytest-asyncio",
    "httpx",
    "email-validator",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
asyncio_mode = "auto"
